[REFACTORING]: renamed VitalityTable components
