+++ v6y-apps/front-bo/src/commons/components/VitalityBaseTable.tsx
@@ -1,0 +1,13 @@
+import { Table } from 'antd';
+import { AnyObject } from 'antd/es/_util/type';
+import React from 'react';
+
+const VitalityTable = ({
+    dataSource,
+    columns,
+}: {
+    dataSource: AnyObject[];
+    columns: AnyObject[];
+}) => <Table dataSource={dataSource} columns={columns} rowKey="key" />;
+
+export default VitalityTable;
