+++ v6y-apps/front-bo/src/commons/components/VitalityTable.tsx
@@ -1,12 +1,33 @@
-import { Table } from 'antd';
-import { AnyObject } from 'antd/es/_util/type';
+import React from 'react';
 
-const VitalityTable = ({
+import { buildCommonTableColumns, buildCommonTableDataSource } from '../config/VitalityTableConfig';
+import VitalityBaseTable from './VitalityBaseTable';
+
+interface RenderVitalityTableProps {
+    dataSource: unknown[];
+    columnKeys: string[];
+    columnOptions: {
+        enableEdit: boolean;
+        enableShow: boolean;
+        enableDelete: boolean;
+        deleteMetaQuery?: {
+            gqlMutation: string;
+            operation: string;
+        };
+    };
+}
+
+const VitalityTable: React.FC<RenderVitalityTableProps> = ({
     dataSource,
-    columns,
-}: {
-    dataSource: AnyObject[];
-    columns: AnyObject[];
-}) => <Table dataSource={dataSource} columns={columns} rowKey="key" />;
+    columnKeys,
+    columnOptions,
+}) => {
+    return (
+        <VitalityBaseTable
+            dataSource={buildCommonTableDataSource(dataSource)}
+            columns={buildCommonTableColumns(dataSource, columnKeys, columnOptions)}
+        />
+    );
+};
 
 export default VitalityTable;
